+++ collector-bitmovin-player-v1/src/test/java/com/bitmovin/analytics/bitmovin/player/BitmovinSdkAdapterTest.kt
@@ -4,13 +4,16 @@ import com.bitmovin.analytics.stateMachines.PlayerStateMachine
 import com.bitmovin.analytics.stateMachines.PlayerStates
 import com.bitmovin.player.BitmovinPlayer
 import com.bitmovin.player.api.event.data.AudioPlaybackQualityChangedEvent
+import com.bitmovin.player.api.event.data.PausedEvent
 import com.bitmovin.player.api.event.listener.EventListener
 import com.bitmovin.player.api.event.listener.OnAudioPlaybackQualityChangedListener
+import com.bitmovin.player.api.event.listener.OnPausedListener
 import com.bitmovin.player.config.quality.AudioQuality
 import io.mockk.every
 import io.mockk.justRun
 import io.mockk.mockk
 import io.mockk.verify
+import io.mockk.clearMocks
 import org.junit.Before
 import org.junit.Test
 
@@ -39,6 +42,45 @@ class BitmovinSdkAdapterTest {
         assert(event != null)
         event!!.onAudioPlaybackQualityChanged(getAudioPlaybackQualityChangedEvent(200, 200))
         verify(exactly = 0) { stateMachine.transitionState(PlayerStates.QUALITYCHANGE, any()) }
+    }
+
+    @Test
+    fun `playerStateMachine does not transition to pause if player is transitioning to ads`() {
+        // arrange
+        val event = getListenerWithType<OnPausedListener>()
+        every { fakePlayer.currentTime } returns 0.0
+        every { fakePlayer.isAd } returns true
+        stateMachine.transitionState(PlayerStates.STARTUP, 0)
+        stateMachine.transitionState(PlayerStates.PLAYING, 0)
+        clearMocks(stateMachine)
+
+        // act
+        assert(event != null)
+        event!!.onPaused(PausedEvent(5.0))
+
+        // asset
+        verify(inverse = true) { stateMachine.pause(any()) }
+        verify(exactly = 1) { stateMachine.startAd(5000) } // 5000 because we convert to ms
+    }
+
+    @Test
+    fun `playerStateMachine transitions to pause if no ads are being played`() {
+        // arrange
+        val event = getListenerWithType<OnPausedListener>()
+        every { fakePlayer.currentTime } returns 0.0
+        every { fakePlayer.isAd } returns false
+        stateMachine.transitionState(PlayerStates.STARTUP, 0)
+        stateMachine.transitionState(PlayerStates.PLAYING, 0)
+        clearMocks(stateMachine)
+
+        // act
+        adapter.init()
+        assert(event != null)
+        event!!.onPaused(PausedEvent(5.0))
+
+        // asset
+        verify(exactly = 1) { stateMachine.pause(any()) }
+        verify(inverse = true) { stateMachine.startAd(5000) } // 5000 because we convert to ms
     }
 
     private fun getAudioPlaybackQualityChangedEvent(oldBitrate: Int, newBitrate: Int): AudioPlaybackQualityChangedEvent {